Added SkillId enum
Used instead of int for all instances where a skill id is used. Makes it easier to reference specific skills and understand what skills portions of the code are for, since you don't need to remember ids.

+++ src/ChannelServer/Network/Helpers/SkillHelper.cs
@@ -20,7 +20,7 @@ namespace Melia.Channel.Network.Helpers
 			var propertiesSize = skill.Properties.Size;
 
 			packet.PutLong(skill.ObjectId);
-			packet.PutInt(skill.Id);
+			packet.PutInt((int)skill.Id);
 			packet.PutShort(propertiesSize);
 			packet.PutEmptyBin(2); // alignment
 			packet.PutInt(0);
